Move components folder to ui folder
