+++ src/battle/components/BattleCard/BattleCard.tsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router";
-import Button from "../../../components/Button/Button";
+import Button from "../../../ui/components/Button/Button";
 import useBattles from "../../hooks/useBattles";
 import type { Battle } from "../../../types";
 import "./BattleCard.css";
