+++ src/ui/components/Navigator/Navigator.css
