+++ src/ui/components/Paginator/Paginator.css
